Split archive.rs into one-pub-per-file module structure
Refactors src/archive.rs into src/archive/ directory following the
project's "one public item per file" convention:

- core.rs: Archive<M> struct, ArchiveReader/ArchiveWriter type aliases
- archive_read.rs: ArchiveRead trait definition
- archive_write.rs: ArchiveWrite trait definition
- reader.rs: impl blocks for Archive<MappedArchive>
- writer.rs: impl blocks for Archive<MappedArchiveMut>, tests
- mod.rs: re-exports, CdEntry, parse_cd_entries, path_to_string helpers

The core.rs file is named to avoid clippy::module_inception (can't have
archive/archive.rs).
