+++ src/archive/writer.rs
@@ -1,531 +1,21 @@
-//! Unified archive access using memory-mapped I/O.
-//!
-//! This module provides a generic `Archive<M>` struct that works with both
-//! read-only and read-write memory maps. The `ArchiveRead` and `ArchiveWrite`
-//! traits define the available operations for each mode.
-//!
-//! # Type Aliases
-//!
-//! - [`ArchiveReader`] = `Archive<MappedArchive>` - Read-only access
-//! - [`ArchiveWriter`] = `Archive<MappedArchiveMut>` - Read-write access
+//! Read-write archive implementation.
 
+use super::{Archive, ArchiveRead, ArchiveWrite, CdEntry, parse_cd_entries, path_to_string};
 use crate::{
     ArchivePath, BaleEocd, BaleError, CentralDirectoryHeader, DosDateTime, Eocd, LocalFileHeader,
-    MappedArchive, MappedArchiveMut, Trailer, Zip64Eocd,
+    MappedArchiveMut, Trailer, Zip64Eocd,
 };
 use std::collections::HashSet;
 use std::fs::File;
 use std::io::Read;
 use std::path::Path;
-use zerocopy::{FromBytes, IntoBytes};
+use zerocopy::IntoBytes;
 
 /// Pre-allocated zero buffer for padding (avoids allocation for common cases).
 ///
 /// Sized to match the default alignment (4096 bytes). For larger alignments,
 /// padding is written in chunks from this buffer.
 static ZERO_PAD: [u8; 4096] = [0u8; 4096];
-
-/// An in-memory Central Directory entry.
-///
-/// Holds the header data and null-padded path for an entry.
-struct CdEntry {
-    /// The Central Directory header.
-    header: CentralDirectoryHeader,
-    /// The null-padded path (length = path_size).
-    path: Vec<u8>,
-}
-
-/// A bale archive with generic memory-map backing.
-///
-/// This struct provides access to archive contents through memory-mapped I/O.
-/// The type parameter `M` determines whether the archive is read-only
-/// ([`MappedArchive`]) or read-write ([`MappedArchiveMut`]).
-///
-/// # Example
-///
-/// ```ignore
-/// use bale::{ArchiveReader, ArchiveRead};
-///
-/// let reader = ArchiveReader::open("archive.bale")?;
-/// for (header, path) in reader.iter_entries() {
-///     let data = reader.read_data(header)?;
-///     // process data...
-/// }
-/// ```
-pub struct Archive<M> {
-    /// The memory-mapped archive file.
-    mmap: M,
-    /// In-memory Central Directory entries.
-    entries: Vec<CdEntry>,
-    /// Archive configuration from the BaleEocd.
-    bale_eocd: BaleEocd,
-    /// Current write offset (end of file data). Only used for writers.
-    write_offset: usize,
-    /// Whether the archive has been modified since the last sync. Only used for writers.
-    dirty: bool,
-}
-
-/// Read-only archive type alias.
-pub type ArchiveReader = Archive<MappedArchive>;
-
-/// Read-write archive type alias.
-pub type ArchiveWriter = Archive<MappedArchiveMut>;
-
-/// Read operations for archives.
-///
-/// This trait is implemented for all `Archive<M>` where `M` provides byte access.
-pub trait ArchiveRead {
-    /// Returns the number of entries in the archive.
-    fn entry_count(&self) -> usize;
-
-    /// Returns the configured path size for this archive.
-    fn path_size(&self) -> usize;
-
-    /// Returns the configured alignment for this archive.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `alignment_pow2` is invalid. This cannot happen for archives
-    /// opened via [`open()`](Archive::open) since validation occurs on construction.
-    fn alignment(&self) -> u32;
-
-    /// Returns the path for the entry at the given index as a zero-copy `ArchivePath`.
-    ///
-    /// The returned path borrows directly from the in-memory entries.
-    /// Returns `None` if the index is out of bounds.
-    fn get_path(&self, index: usize) -> Option<ArchivePath<'_>>;
-
-    /// Returns an iterator over all Central Directory entries.
-    ///
-    /// Each item is a tuple of (header, path_bytes) where path_bytes is the
-    /// null-padded path from the CD entry.
-    fn iter_entries(&self) -> impl Iterator<Item = (&CentralDirectoryHeader, &[u8])>;
-
-    /// Finds an entry by path using linear scan.
-    ///
-    /// Returns the last matching entry. Bale uses append-only shadowing: when
-    /// a file is updated, the new version is appended and the old version
-    /// remains but is "shadowed". The last occurrence is the current version.
-    ///
-    /// The path comparison is byte-exact against the null-padded path.
-    fn find_entry(&self, path: &str) -> Option<&CentralDirectoryHeader>;
-
-    /// Returns a zero-copy slice of the file data for the given entry.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the entry's offset or size is invalid.
-    fn read_data(&self, entry: &CentralDirectoryHeader) -> Result<&[u8], BaleError>;
-
-    /// Returns a reference to the BaleEocd.
-    fn bale_eocd(&self) -> &BaleEocd;
-
-    /// Verifies the CRC-32 checksum for an entry.
-    ///
-    /// Reads the entry data and computes its CRC-32, comparing against the
-    /// stored value in the Central Directory header.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - The entry data cannot be read
-    /// - The computed CRC does not match the stored CRC
-    fn verify_crc(&self, entry: &CentralDirectoryHeader) -> Result<(), BaleError>;
-
-    /// Checks if the Central Directory is sorted by path bytes.
-    ///
-    /// A sorted CD enables binary search for entry lookup. Archives created
-    /// by `compact` are always sorted.
-    fn is_sorted(&self) -> bool;
-
-    /// Returns a list of duplicate paths in the archive.
-    ///
-    /// Duplicate paths occur when the same path appears multiple times in the
-    /// Central Directory (shadowing). Returns the paths that have duplicates,
-    /// not the total count of duplicates.
-    fn find_duplicates(&self) -> Vec<String>;
-
-    /// Checks if the archive contains orphaned data.
-    ///
-    /// Orphaned data exists when there are gaps between entries or between
-    /// the last entry and the Central Directory. This can occur after
-    /// deletions or when entries are shadowed.
-    fn has_orphaned_data(&self) -> bool;
-}
-
-/// Write operations for archives.
-///
-/// This trait is only implemented for `Archive<MappedArchiveMut>`.
-pub trait ArchiveWrite: ArchiveRead {
-    /// Adds an entry from raw data.
-    ///
-    /// If an entry with the same path already exists, the new entry shadows it.
-    /// The old data remains in the archive (orphaned) until compact.
-    ///
-    /// # Arguments
-    ///
-    /// * `path` - Archive path for the entry
-    /// * `data` - File contents
-    /// * `mode` - Unix file permissions (e.g., 0o644)
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - The path exceeds the archive's path_size
-    /// - The data size exceeds 4GB (ZIP format limitation)
-    /// - The archive offset would exceed 4GB (ZIP format limitation)
-    /// - Writing to the archive fails
-    fn add_entry(&mut self, path: &str, data: &[u8], mode: u32) -> Result<(), BaleError>;
-
-    /// Adds a file from the filesystem to the archive.
-    ///
-    /// # Memory usage
-    ///
-    /// This method reads the entire file into memory before writing to the
-    /// archive. For very large files, consider using [`add_entry()`](Self::add_entry)
-    /// with a streaming approach, or ensure sufficient memory is available.
-    ///
-    /// # Arguments
-    ///
-    /// * `src` - Path to the source file
-    /// * `archive_path` - Path within the archive
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - The source file cannot be read
-    /// - The archive path exceeds path_size
-    /// - The file size exceeds 4GB (ZIP format limitation)
-    /// - Writing to the archive fails
-    fn add_file(&mut self, src: impl AsRef<Path>, archive_path: &str) -> Result<(), BaleError>;
-
-    /// Deletes all entries matching a path.
-    ///
-    /// Removes all matching entries from the Central Directory. If duplicate
-    /// entries exist (from shadowing), all are removed. The file data remains
-    /// in the archive (orphaned) until a compact operation.
-    ///
-    /// Returns `true` if any entries were deleted, `false` if none matched.
-    fn delete(&mut self, path: &str) -> bool;
-
-    /// Flushes all changes to disk.
-    ///
-    /// Rewrites the Central Directory and full trailer (ZIP64 EOCD, ZIP64 EOCD
-    /// Locator, EOCD, and BaleEocd). The CD starts at an aligned offset for
-    /// efficient mmap access. The file is truncated to the logical size.
-    ///
-    /// If no changes have been made since the last sync, this is a no-op.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if writing or syncing fails.
-    fn sync(&mut self) -> Result<(), BaleError>;
-}
-
-// ============================================================================
-// Shared implementation for parsing archives
-// ============================================================================
-
-/// Parses Central Directory entries from archive bytes.
-///
-/// # Arguments
-///
-/// * `bytes` - The full archive bytes
-/// * `cd_offset` - Offset to the start of the Central Directory
-/// * `entry_count` - Number of entries in the CD
-/// * `path_size` - Size of each path field
-///
-/// # Errors
-///
-/// Returns an error if any CD entry is malformed or extends beyond the archive.
-fn parse_cd_entries(
-    bytes: &[u8],
-    cd_offset: usize,
-    entry_count: usize,
-    path_size: usize,
-) -> Result<Vec<CdEntry>, BaleError> {
-    let stride = CentralDirectoryHeader::stride(path_size);
-    let mut entries = Vec::with_capacity(entry_count);
-
-    for i in 0..entry_count {
-        let entry_start = cd_offset + i * stride;
-        let entry_end = entry_start + stride;
-        if entry_end > bytes.len() {
-            return Err(BaleError::Corrupted(format!(
-                "CD entry {i} extends beyond archive"
-            )));
-        }
-
-        let entry_bytes = &bytes[entry_start..entry_end];
-        let header =
-            CentralDirectoryHeader::ref_from_bytes(&entry_bytes[..CentralDirectoryHeader::SIZE])
-                .map_err(|e| BaleError::Corrupted(format!("invalid CD entry {i}: {e}")))?;
-
-        entries.push(CdEntry {
-            header: *header,
-            path: entry_bytes[CentralDirectoryHeader::SIZE..].to_vec(),
-        });
-    }
-
-    Ok(entries)
-}
-
-// ============================================================================
-// ArchiveReader implementation
-// ============================================================================
-
-impl Archive<MappedArchive> {
-    /// Opens an archive for reading.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - The file cannot be opened or memory-mapped
-    /// - The archive is too small to contain a valid trailer
-    /// - Any trailer signature is invalid (ZIP64 EOCD, Locator, EOCD, or BaleEocd)
-    pub fn open(path: impl AsRef<Path>) -> Result<Self, BaleError> {
-        let mmap = MappedArchive::open(path)?;
-        let bytes = mmap.as_bytes();
-
-        // Parse and validate trailer.
-        let trailer = Trailer::from_archive_bytes(bytes)?;
-        let bale_eocd = trailer.bale_eocd;
-        let path_size = trailer.path_size() as usize;
-        let cd_offset = trailer.cd_offset() as usize;
-        let entry_count = trailer.entry_count() as usize;
-
-        // Parse Central Directory entries.
-        let entries = parse_cd_entries(bytes, cd_offset, entry_count, path_size)?;
-
-        Ok(Self {
-            mmap,
-            entries,
-            bale_eocd,
-            write_offset: 0,
-            dirty: false,
-        })
-    }
-
-    /// Returns a reference to the parsed trailer.
-    ///
-    /// Note: This reconstructs the trailer from stored components. For most
-    /// uses, prefer the individual accessor methods like [`bale_eocd()`](Self::bale_eocd).
-    #[must_use]
-    pub fn trailer(&self) -> Trailer {
-        // Reconstruct trailer from stored data.
-        // CD offset is computed from entries.
-        let cd_offset = self.compute_cd_offset();
-        let cd_size = self.entries.len() * CentralDirectoryHeader::stride(self.path_size());
-        let zip64_eocd_offset = cd_offset + cd_size;
-
-        Trailer::new(
-            self.entries.len() as u64,
-            cd_size as u64,
-            cd_offset as u64,
-            zip64_eocd_offset as u64,
-            self.bale_eocd,
-        )
-    }
-
-    /// Returns a reference to the ZIP64 EOCD.
-    #[must_use]
-    pub fn zip64_eocd(&self) -> Zip64Eocd {
-        self.trailer().zip64_eocd
-    }
-
-    /// Returns a reference to the EOCD.
-    #[must_use]
-    pub fn eocd(&self) -> Eocd {
-        self.trailer().eocd
-    }
-
-    /// Computes the CD offset from entries.
-    fn compute_cd_offset(&self) -> usize {
-        if self.entries.is_empty() {
-            return 0;
-        }
-
-        // Find the highest local_header_offset + entry size.
-        let path_size = self.path_size();
-        let local_stride = LocalFileHeader::stride(path_size);
-        let alignment = self.alignment() as usize;
-
-        let mut max_end: usize = 0;
-        for entry in &self.entries {
-            let local_offset = entry.header.local_header_offset.get() as usize;
-            let data_size = entry.header.uncompressed_size.get() as usize;
-            let unaligned_size = local_stride + data_size;
-            let aligned_size = unaligned_size.div_ceil(alignment) * alignment;
-            let entry_end = local_offset + aligned_size;
-            max_end = max_end.max(entry_end);
-        }
-
-        max_end
-    }
-}
-
-impl ArchiveRead for Archive<MappedArchive> {
-    fn entry_count(&self) -> usize {
-        self.entries.len()
-    }
-
-    fn path_size(&self) -> usize {
-        self.bale_eocd.path_size() as usize
-    }
-
-    fn alignment(&self) -> u32 {
-        self.bale_eocd
-            .alignment()
-            .expect("validated on construction")
-    }
-
-    fn get_path(&self, index: usize) -> Option<ArchivePath<'_>> {
-        let entry = self.entries.get(index)?;
-        let path_bytes = &entry.path;
-
-        // Trim null padding for the ArchivePath.
-        let end = path_bytes
-            .iter()
-            .position(|&b| b == 0)
-            .unwrap_or(path_bytes.len());
-
-        Some(ArchivePath::from_bytes(&path_bytes[..end]))
-    }
-
-    fn iter_entries(&self) -> impl Iterator<Item = (&CentralDirectoryHeader, &[u8])> {
-        self.entries
-            .iter()
-            .map(|entry| (&entry.header, entry.path.as_slice()))
-    }
-
-    fn find_entry(&self, path: &str) -> Option<&CentralDirectoryHeader> {
-        let path_bytes = path.as_bytes();
-        let path_size = self.path_size();
-
-        if path_bytes.len() > path_size {
-            return None;
-        }
-
-        let mut result = None;
-        for entry in &self.entries {
-            if entry.path.starts_with(path_bytes)
-                && entry.path[path_bytes.len()..].iter().all(|&b| b == 0)
-            {
-                result = Some(&entry.header);
-            }
-        }
-        result
-    }
-
-    fn read_data(&self, entry: &CentralDirectoryHeader) -> Result<&[u8], BaleError> {
-        let bytes = self.mmap.as_bytes();
-        let local_offset = entry.local_header_offset.get() as usize;
-        let path_size = self.path_size();
-        let data_size = entry.uncompressed_size.get() as usize;
-
-        let local_stride = LocalFileHeader::stride(path_size);
-        let data_start = local_offset
-            .checked_add(local_stride)
-            .ok_or_else(|| BaleError::Corrupted("offset overflow".to_string()))?;
-        let data_end = data_start
-            .checked_add(data_size)
-            .ok_or_else(|| BaleError::Corrupted("size overflow".to_string()))?;
-
-        if data_end > bytes.len() {
-            return Err(BaleError::Corrupted(format!(
-                "entry data extends beyond archive: offset={local_offset}, size={data_size}"
-            )));
-        }
-
-        Ok(&bytes[data_start..data_end])
-    }
-
-    fn bale_eocd(&self) -> &BaleEocd {
-        &self.bale_eocd
-    }
-
-    fn verify_crc(&self, entry: &CentralDirectoryHeader) -> Result<(), BaleError> {
-        let data = self.read_data(entry)?;
-        let computed = crc32fast::hash(data);
-        let stored = entry.crc32.get();
-
-        if computed != stored {
-            return Err(BaleError::Corrupted(format!(
-                "CRC mismatch: expected {:08x}, got {:08x}",
-                stored, computed
-            )));
-        }
-
-        Ok(())
-    }
-
-    fn is_sorted(&self) -> bool {
-        let mut prev: Option<&[u8]> = None;
-        for entry in &self.entries {
-            if let Some(p) = prev
-                && p > entry.path.as_slice()
-            {
-                return false;
-            }
-            prev = Some(&entry.path);
-        }
-        true
-    }
-
-    fn find_duplicates(&self) -> Vec<String> {
-        let mut seen: HashSet<&[u8]> = HashSet::new();
-        let mut duplicate_set: HashSet<&[u8]> = HashSet::new();
-
-        for entry in &self.entries {
-            if !seen.insert(&entry.path) {
-                duplicate_set.insert(&entry.path);
-            }
-        }
-
-        duplicate_set.into_iter().map(path_to_string).collect()
-    }
-
-    fn has_orphaned_data(&self) -> bool {
-        if self.entries.is_empty() {
-            return false;
-        }
-
-        // Sort entries by local_header_offset.
-        let mut sorted_entries: Vec<_> = self.entries.iter().collect();
-        sorted_entries.sort_by_key(|e| e.header.local_header_offset.get());
-
-        let path_size = self.path_size();
-        let alignment = self.alignment() as usize;
-        let local_header_stride = LocalFileHeader::stride(path_size);
-        let cd_offset = self.compute_cd_offset();
-
-        let mut expected_offset: usize = 0;
-
-        for entry in &sorted_entries {
-            let local_offset = entry.header.local_header_offset.get() as usize;
-            let data_size = entry.header.uncompressed_size.get() as usize;
-
-            if local_offset != expected_offset {
-                return true;
-            }
-
-            let Some(entry_size) = local_header_stride.checked_add(data_size) else {
-                return true;
-            };
-            let aligned_size = entry_size.div_ceil(alignment).saturating_mul(alignment);
-            let Some(next_offset) = local_offset.checked_add(aligned_size) else {
-                return true;
-            };
-            expected_offset = next_offset;
-        }
-
-        expected_offset != cd_offset
-    }
-}
-
-// ============================================================================
-// ArchiveWriter implementation
-// ============================================================================
 
 impl Archive<MappedArchiveMut> {
     /// Creates a new empty archive at the given path.
@@ -952,26 +442,10 @@ impl ArchiveWrite for Archive<MappedArchiveMut> {
     }
 }
 
-// ============================================================================
-// Helper functions
-// ============================================================================
-
-/// Converts a null-padded path to a string.
-fn path_to_string(path_bytes: &[u8]) -> String {
-    let end = path_bytes
-        .iter()
-        .position(|&b| b == 0)
-        .unwrap_or(path_bytes.len());
-    String::from_utf8_lossy(&path_bytes[..end]).to_string()
-}
-
-// ============================================================================
-// Tests
-// ============================================================================
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::{ArchiveReader, ArchiveWriter};
     use std::io::Write;
     use tempfile::{NamedTempFile, TempDir};
